// src/library/enums/elastic-search-keys.enum.ts
export enum elasticSearchKey {
    COMPANYID = 'COMPANYID',
    COMPANYTYPEID = 'COMPANYTYPEID',
    COMPANYSTATUSTYPEID = 'COMPANYSTATUSTYPEID',
    SIMPLEINDUSTRYID = 'SIMPLEINDUSTRYID',
    COMPANYNAME = 'COMPANYNAME',
    CITY = 'CITY',
    PRICINGDATE = 'PRICINGDATE',
    SEGMENTDESCRIPTION = 'SEGMENTDESCRIPTION',
    CALENDARYEAR = 'CALENDARYEAR',
    CALENDARQUARTER = 'CALENDARQUARTER',
    DATAITEMID = 'DATAITEMID'
}